+++ EasySoccer.WebApi/EasySoccer.BLL/SoccerPitchPlanBLL.cs
@@ -35,7 +35,7 @@ namespace EasySoccer.BLL
             return soccerPitchPlan;
         }
 
-        public Task<List<SoccerPitchPlan>> GetAsync(int companyId, int page, int pageSize)
+        public Task<List<SoccerPitchPlan>> GetAsync(long companyId, int page, int pageSize)
         {
             return _soccerPitchPlanRepository.GetAsync(companyId, page, pageSize);
         }
