+++ EasySoccer.WebApi/EasySoccer.DAL.Infra/Repositories/ISoccerPitchPlanRepository.cs
@@ -6,7 +6,7 @@ namespace EasySoccer.DAL.Infra.Repositories
 {
     public interface ISoccerPitchPlanRepository: IRepositoryBase
     {
-        Task<List<SoccerPitchPlan>> GetAsync(int companyId, int page, int pageSize);
+        Task<List<SoccerPitchPlan>> GetAsync(long companyId, int page, int pageSize);
         Task<SoccerPitchPlan> GetAsync(int id);
     }
 }
